Add docs; discover bug
If the maxSlop is set very high (higher than the purchase price),
the script *should* converge to a one-step purchase that spends
all the money at once. The algorithm is instead doing one purchase
at the lowest price and then punting.

It looks like the logic actually needs to run in reverse:
 - A one-step purchase is the least precise repro of the motif.
 - Adding more steps should make it more accurate.

So we need to start from the maximum purchase, and then divide out
the remaining money across the symbols we *haven't* purchased yet.
We'll have to record the least-precise purchase, and then refine
it step by step to the most-precise.

This will be a very different algorithm.