Autoscaling; step numbering
Because not all the stocks are available in slices, add an autoscale
feature to total up the percentages (which will be less than 100%
because of missing stocks), scale that to 100%, and scale back the
investment amount so they still match up in total amount spent.

The step numbering wasn't incrementing; added an increment.